Combined watch with convert. Bug fixes. ```
