+++ convert/package.json
@@ -1,0 +1,25 @@
+{
+  "dependencies": {
+    "chokidar": "^3.5.3",
+    "clipboardy": "^3.0.0",
+    "fluent-ffmpeg": "^2.1.2",
+    "typescript": "^5.2.2"
+  },
+  "name": "convertogg2mp3",
+  "version": "1.0.0",
+  "main": "index.js",
+  "type": "module",
+  "scripts": {
+    "start": "tsc && node dist/index.js"
+  },
+  "keywords": [],
+  "author": "aramrw",
+  "license": "ISC",
+  "description": "Script to watch Anki's media folder and convert .ogg files to .mp3.",
+  "devDependencies": {
+    "@types/chokidar": "^2.1.3",
+    "@types/clipboardy": "^2.0.1",
+    "@types/fluent-ffmpeg": "^2.1.22",
+    "@types/node": "^20.6.2"
+  }
+}
