+++ convertogg2mp3/run.bat
@@ -1,0 +1,2 @@
+@echo off
+start /min powershell.exe -ExecutionPolicy Bypass -File "%~dp0convertogg2mp3.ps1"
