--- package.json
@@ -1,22 +1,0 @@
-{
-  "dependencies": {
-    "chokidar": "^3.5.3",
-    "fluent-ffmpeg": "^2.1.2",
-    "typescript": "^5.2.2"
-  },
-  "name": "convertogg2mp3",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "node dist/index.js"
-  },
-  "keywords": [],
-  "author": "aramrw",
-  "license": "ISC",
-  "description": "Script to watch Anki's media folder and convert .ogg files to .mp3.",
-  "devDependencies": {
-    "@types/chokidar": "^2.1.3",
-    "@types/fluent-ffmpeg": "^2.1.22",
-    "@types/node": "^20.6.2"
-  }
-}
