--- run.bat
@@ -1,2 +1,0 @@
-@echo off
-start /min powershell.exe -ExecutionPolicy Bypass -File "%~dp0convertogg2mp3.ps1"
