--- src/mining.ts
@@ -1,2 +1,0 @@
-import * as chokidar from 'chokidar';
-
