+++ watch-mining-dev/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "watch-mining",
+  "version": "1.0.0",
+  "description": "",
+  "main": "watch-mining.js",
+  "type": "commonjs",
+  "scripts": {
+    "build": "babel watch-mining.js -d dist",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "chokidar": "^3.5.3",
+    "clipboardy": "^3.0.0",
+    "node-clipboardy": "^1.0.3",
+    "pkg": "^5.8.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.22.20",
+    "@babel/preset-env": "^7.22.20",
+    "esbuild": "^0.19.3"
+  }
+}
